+++ MESysWin/src/FuzzyVariable.cs
@@ -9,23 +9,23 @@ namespace MESysWin.src
 {
     class FuzzyVariable
     {
-        public int ID { get; set; }
+        public long ID { get; set; }
 
-        public int IdSymptom { get; set; }
+        public long IdSymptom { get; set; }
 
         public string Name { get; set; }
 
         public TypeMFuncEnum Type { get; set; }
 
-        public int IdTriangulare { get; set; }
+        public long IdTriangulare { get; set; }
 
-        public int IdTrapezoidal { get; set; }
+        public long IdTrapezoidal { get; set; }
 
-        public int IdGaussian { get; set; }
+        public long IdGaussian { get; set; }
 
         public Color Color { get; set; }
 
-        public FuzzyVariable(int id_symp, string name, Color color)
+        public FuzzyVariable(long id_symp, string name, Color color)
         {
             IdSymptom = id_symp;
             Name = name;
