+++ MESysWin/src/TypeMFunc.cs
@@ -10,13 +10,13 @@ namespace MESysWin.src
 
     class TypeMFunc
     {
-        public int ID { get; set; }
+        public long ID { get; set; }
 
         public string Name { get; set; }
 
         public string Description { get; set; }
 
-        public TypeMFunc(int id, string name, string description)
+        public TypeMFunc(long id, string name, string description)
         {
             ID = id;
             Name = name;
